requirements.txt
bcrypt==3.2.0
blinker==1.4
cachetools==4.1.1
certifi==2020.11.8
cffi==1.14.3
chardet==3.0.4
Click==7.0
dnspython==2.0.0
email-validator==1.1.1
Flask==1.1.1
Flask-Bcrypt==0.7.1
Flask-DebugToolbar==0.11.0
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
google-api-core==1.23.0
google-api-python-client==1.12.8
google-auth==1.23.0
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.2
googleapis-common-protos==1.52.0
gunicorn==20.0.4
httplib2==0.18.1
idna==2.10
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
oauthlib==3.1.0
protobuf==3.14.0
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pytz==2020.4
requests==2.25.0
requests-oauthlib==1.3.0
rsa==4.6
six==1.15.0
SQLAlchemy==1.3.9
uritemplate==3.0.1
urllib3==1.26.2
Werkzeug==0.16.0
WTForms==2.3.3
